chore: Integrate learning parameters with new training configuration
