+++ web_ui/src/pages/project-details/components/project-models/train-model-dialog/advanced-settings/ui/parameters.component.tsx
@@ -3,14 +3,10 @@
 
 import { FC, ReactNode } from 'react';
 
-import { Flex, Grid, minmax, Switch, Text, ToggleButtons, View } from '@geti/ui';
+import { Grid, minmax, Text, ToggleButtons, View } from '@geti/ui';
 import { isFunction } from 'lodash-es';
 
-import {
-    BoolParameter,
-    ConfigurationParameter,
-} from '../../../../../../../core/configurable-parameters/services/configuration.interface';
-import { isBoolParameter } from '../../../../../../../core/configurable-parameters/utils';
+import { ConfigurationParameter } from '../../../../../../../core/configurable-parameters/services/configuration.interface';
 import { BooleanParameter } from './boolean-parameter.component';
 import { NumberParameterField } from './number-parameter-field.component';
 import { ResetButton } from './reset-button.component';
@@ -18,7 +14,7 @@ import { Tooltip } from './tooltip.component';
 
 interface ParametersProps {
     parameters: ConfigurationParameter[];
-    onChange: () => void;
+    onChange: (value: string | boolean | number) => void;
 }
 
 const ParameterTooltip: FC<{ text: string }> = ({ text }) => {
@@ -27,7 +23,7 @@ const ParameterTooltip: FC<{ text: string }> = ({ text }) => {
 
 interface ParameterProps {
     parameter: ConfigurationParameter;
-    onChange: () => void;
+    onChange: (value: string | boolean | number) => void;
     isDisabled?: boolean;
 }
 
@@ -51,7 +47,7 @@ const ParameterLayout: FC<ParameterLayoutProps> = ({ header, children, descripti
     );
 };
 
-export const ParameterField: FC<ParameterProps> = ({ parameter, onChange, isDisabled }) => {
+const ParameterField: FC<ParameterProps> = ({ parameter, onChange, isDisabled }) => {
     if (parameter.type === 'enum') {
         return (
             <ToggleButtons
@@ -71,54 +67,59 @@ export const ParameterField: FC<ParameterProps> = ({ parameter, onChange, isDisa
                 maxValue={parameter.maxValue}
                 onChange={onChange}
                 type={parameter.type}
+                isDisabled={isDisabled}
             />
         );
     }
 
     if (parameter.type === 'bool') {
-        return <BooleanParameter value={parameter.value} header={parameter.name} onChange={() => {}} />;
+        return (
+            <BooleanParameter
+                value={parameter.value}
+                header={parameter.name}
+                onChange={onChange}
+                isDisabled={isDisabled}
+            />
+        );
     }
 };
 
-export const Parameter: FC<ParameterProps> = ({ parameter, onChange }) => {
+export const Parameter = ({ parameter, onChange, isDisabled }: ParameterProps) => {
+    const handleReset = () => {
+        onChange(parameter.defaultValue);
+    };
+
     return (
-        <ParameterLayout header={parameter.name} description={parameter.description} onReset={() => {}}>
-            <ParameterField parameter={parameter} onChange={onChange} />
+        <ParameterLayout header={parameter.name} description={parameter.description} onReset={handleReset}>
+            <ParameterField parameter={parameter} onChange={onChange} isDisabled={isDisabled} />
         </ParameterLayout>
     );
 };
 
+Parameter.Layout = ParameterLayout;
+Parameter.Field = ParameterField;
+
 interface ParametersListProps {
     parameters: ConfigurationParameter[];
-    onChange: () => void;
+    onChange: (value: string | boolean | number) => void;
 }
 
 const ParametersList = ({ parameters, onChange }: ParametersListProps) => {
-    if (parameters.length === 2 && parameters[0].key === 'enable' && isBoolParameter(parameters[0])) {
-        const enableParameter = parameters[0] as BoolParameter;
-        const configParameter = parameters[1];
-
-        return (
-            <ParameterLayout header={configParameter.name} description={configParameter.description} onReset={() => {}}>
-                <Flex gap={'size-100'}>
-                    <Switch
-                        isSelected={enableParameter.value}
-                        onChange={() => {}}
-                        aria-label={`Toggle ${configParameter.name}`}
-                    />
-                    <ParameterField parameter={configParameter} onChange={onChange} />
-                </Flex>
-            </ParameterLayout>
-        );
-    }
-
     return parameters.map((parameter) => <Parameter key={parameter.name} parameter={parameter} onChange={onChange} />);
 };
 
-export const Parameters: FC<ParametersProps> = ({ parameters, onChange }) => {
+export const Parameters = ({ parameters, onChange }: ParametersProps) => {
     return (
         <Grid columns={['size-3000', minmax('size-3400', '1fr'), 'size-400']} gap={'size-300'} alignItems={'center'}>
             <ParametersList parameters={parameters} onChange={onChange} />
+        </Grid>
+    );
+};
+
+Parameters.Container = ({ children }: { children: ReactNode }) => {
+    return (
+        <Grid columns={['size-3000', minmax('size-3400', '1fr'), 'size-400']} gap={'size-300'} alignItems={'center'}>
+            {children}
         </Grid>
     );
 };
